Refine Mason SDK client structure
